+++ test/spec/utilities_test.js
@@ -1,6 +1,6 @@
 describe('Cape', function() {
   describe('extend', function() {
-    it('merge the properties of objects into the first.', function() {
+    it('should merge the properties of objects into the first.', function() {
       var obj1 = { a: 1, b: 2, c: 3 };
       var obj2 = { a: 4, d: 5 };
       var obj3 = { d: 6, e: 7 };
@@ -15,7 +15,8 @@ describe('Cape', function() {
   })
 
   describe('deepExtend', function() {
-    it('merge the properties of objects into the first recursively.', function() {
+    it('should merge the properties of objects into the first recursively.',
+      function() {
       var obj1 = { a: { b: { c: 1, d: 2 } }, x: { y: 3 } };
       var obj2 = { a: { b: { d: 4 } }, x: { z: 5 } };
       var obj3 = { a: { b: { e: { f: 6 } } }, x: { } };
@@ -26,6 +27,22 @@ describe('Cape', function() {
       expect(obj1.a.b.e.f).to.equal(6);
       expect(obj1.x.y).to.equal(3);
       expect(obj1.x.z).to.equal(5);
+    })
+  })
+
+  describe('merge', function() {
+    it('should merge (but not override) the properties of objects into the first.',
+      function() {
+      var obj1 = { a: 1, b: 2, c: 3 };
+      var obj2 = { a: 4, d: 5 };
+      var obj3 = { d: 6, e: 7 };
+
+      Cape.merge(obj1, obj2, obj3);
+      expect(obj1.a).to.equal(1);
+      expect(obj1.b).to.equal(2);
+      expect(obj1.c).to.equal(3);
+      expect(obj1.d).to.equal(5);
+      expect(obj1.e).to.equal(7);
     })
   })
 })